python_script/pyproject.toml
[project]
name = "python-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
